requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
asyncio==3.4.3
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dotenv==0.9.9
frozenlist==1.5.0
greenlet==3.1.1
idna==3.10
liqpay-python @ git+https://github.com/liqpay/sdk-python@ae287304c5183f525b94e5aa57d8b8a40a3233a0
magic-filter==1.0.12
multidict==6.1.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
SQLAlchemy==2.0.39
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
